Clean up tests for known helpers functionality
The test was originally skipping the template evaluation, and was
instead reyling on doing string matches on the actual template source.
This might break with future versions of handlebars, so it's better
to just test that the behavior is correct - it shouldn't attempt
to require the known helper. We can stub out a handlebars runtime
environment in the test which includes the known helper.